moved templating support into a plugins directory that can optionally be used when one of the templateing libraries is present
