+++ test/Project.toml
@@ -5,17 +5,17 @@ JSON3 = "0f8b85d8-7281-11e9-16c2-39a750bddbf1"
 MIMEs = "6c6e2e6c-3030-632d-7369-2d6c69616d65"
 Mustache = "ffc61752-8dc7-55ee-8c37-f3e9cdd09e70"
 OteraEngine = "b2d7f28f-acd6-4007-8b26-bc27716e5513"
+Pkg = "44cfe95a-1eb2-52ea-b672-e2afdf69b78f"
 RelocatableFolders = "05181044-ff0b-4ac5-8273-598c1e38db00"
+Requires = "ae029012-a4dd-5104-9daa-d747884805df"
 Sockets = "6462fe0b-24de-5631-8697-dd941f90decc"
 StructTypes = "856f2bd8-1eba-4b0a-8007-ebc267875bd4"
 Test = "8dfed614-e22c-5e08-85e1-65c5234f0b40"
 
 [compat]
+julia = "^1.6"
 HTTP = "^1"
 JSON3 = "^1.9"
 MIMEs = "^0.1.4"
+Requires = "^1"
 RelocatableFolders = "^1"
-Mustache = "^1"
-OteraEngine = "^0"
-julia = "^1.6"
-
